Add build constraints to support non-linux systems
